TEST: check that entering Leemur had the expected effect

+++ tests/name-game.sikuli/name-game.py
@@ -24,15 +24,18 @@ def main():
     click("ctp2launch-btn.png")
     if exists("ctp2ctr-bar.png", 100):
         if waitVanish("ctp2progress-bar.png", 100): # control bar appears before progressbar vanishes
-            file = capture(SCREEN.getBounds())
-            if file:
-                f= Finder(file) # http://doc.sikuli.org/finder.html
-                f.find("ctp2ctr-bar.png")
-                if not f.hasNext():
-                    print("Pattern not found in screen shot: " + file)
-                    exit(30)
-                shutil.move(file, bsfn + '.png')
-                exit(0)
+            if exists(Pattern("ctp2leemur.png").similar(0.90), 100):
+                file = capture(SCREEN.getBounds())
+                if file:
+                    f= Finder(file) # http://doc.sikuli.org/finder.html
+                    f.find("ctp2leemur.png")
+                    if not f.hasNext():
+                        print("Pattern not found in screen shot: " + file)
+                        exit(30)
+                    shutil.move(file, bsfn + '.png')
+                    exit(0)
+            else:
+                exit(35)
         else:
             exit(20)
     else:
